+++ sysmanager Maven Webapp/src/main/java/com/yb/service/VipTagService.java
@@ -9,7 +9,7 @@ public interface VipTagService {
 			List<String> age,List<String> type,List<String> coupon,
 			List<String> recentOil,List<String> recentNotOil,List<String> shortOil,
 			Integer index,Integer number);
-	public Double queryTotal(List<String> loyalty,List<String> identity,List<String> gender,
+	public Integer queryTotal(List<String> loyalty,List<String> identity,List<String> gender,
 			List<String> age,List<String> type,List<String> coupon,
 			List<String> recentOil,List<String> recentNotOil,List<String> shortOil);
 }
